% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_reads.R
\name{bin_reads}
\alias{bin_reads}
\title{Count reads in genomic bins}
\usage{
bin_reads(reads, bins)
}
\arguments{
\item{reads}{a data.frame of read positions for a given individual}

\item{bins}{a list (length of which = number of chr) of breakpoints}
}
\value{
To be completed
}
\description{
A function to put every the position of every read listed in a
data.frame into its corresponding bin.
}
\examples{
NULL

}
