% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_events.R
\name{call_events}
\alias{call_events}
\title{Call CNV events from a set of read counts}
\usage{
call_events(read_counts, min_reads, max_reads, seg_threshold, homdel = -2.5,
  hetdel = -0.5, dup = 0.2)
}
\arguments{
\item{read_counts}{a \code{data.frame} of read counts as generated from
function \code{link{read_counts}}.}

\item{min_reads}{a single numeric value. The minimum mean number of reads per
individual that a bin must have to be kept for further processing.}

\item{max_reads}{a single numeric value. The maximum mean number of reads per
individual that a bin can have to be kept for further processing.}

\item{seg_threshold}{a single numeric value. The threshold parameter used
for segmentation. See function \code{\link{segment}} for more details.}

\item{homdel}{a single numeric value. The threshold to be used for calling
homozygous deletions; all segments with a mean log2 ratio less than this
values is considered called as a homozygous deletion.}

\item{hetdel}{a single numeric value. The threshold to be used for calling
hemizygous deletions; all segments with a mean log2 ratio less than this
value and great than or equal to \code{homdel} is called as a hemizygous
deletion.}

\item{dup}{a single numeric value. The threshold to be used for calling
duplications; all segments with a mean log2 ratio larger than this value
is called as a duplication.}
}
\value{
a list of two elements :
- \code{segdels}: a \code{data.frame} containing the CNV events called and
  some metadata about them.
- \code{log2_ratios}: a \code{data.frame} of log2 ratios per bin and per
  sample. The first three columns indicate the genomic position of the
  bin, whereas each remaining column represents the data for a single
  sample (column names correspond to sample names).
}
\description{
This function performs all the steps needed to go from a \code{data.frame}
read counts per bin and per individual to a set of CNVs. The steps involved
are the normalization of the read counts per sample, the removal of bins
deemed uninformative for the purposes of CNV calling, the computation of
log2 ratios and segmentation based on these ratios, and finally CNV calling
based on user-defined log2 ratio thresholds.
}
\examples{
NULL
}
