% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_out.R
\name{filter_out}
\alias{filter_out}
\title{Filter out events located in heterogeneity areas}
\usage{
filter_out(event_data, het_regions, min_overlap = 0)
}
\arguments{
\item{event_data}{a \code{data.frame} of events such as generated by function
\code{\link{getdels}}}

\item{het_regions}{a \code{data.frame} of genomic ranges to be used for
filtering out events located in these regions.}

\item{min_overlap}{a single numeric value between 0 and 1. The minimal
proportion of the length of the event that must overlap with a region
listed in \code{het_sites} for this event to be filtered out. A value of
0 results in an overlap of even a single nucleotide to be removed, whereas
a value of 1 results in only CNVs entirely located in a specified region to
be removed.}
}
\value{
a \code{data.frame} of read counts similar to that given as input,
  but with events removed according to the specified filters.
}
\description{
Filter out events located in heterogeneity areas
}
\examples{
NULL
}
