% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pos.R
\name{get_pos}
\alias{get_pos}
\title{A function to get the positions of the reads given a BAM file}
\usage{
get_pos(filename, minq = 30, strand_diff = FALSE)
}
\arguments{
\item{filename}{character, the name of the bam file}

\item{minq}{integer, the minimum mapping quality for a read to be kept}

\item{strand_diff}{a logical of length 1 (\code{TRUE} or \code{FALSE}).
If \code{TRUE}, reads mapping to the "+" strand are tallied according to
their 3'-most position, whereas reads mapping to the "-" strand are tallied
according to their 5'-most position. If \code{FALSE}, all reads are tallied
according to their 5'-most position. This option should be chosen according
to the expected mean size of the restriction fragments, so as to maximize
the probability that two reads originating from the same restriction
fragment but with differing orientation are counted in the same bin.}
}
\value{
To be completed
}
\description{
Outputs a data.frame containing the chromosome, initial base
position, and mapping quality for every read in the data set. Filters
based on mapping quality (default = 30, but different thresholds
should be tested).
}
\examples{
NULL
}
