% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdels.R
\name{getdels}
\alias{getdels}
\title{Extract putative CNVs from a set of segments}
\usage{
getdels(data, ind, all_segments = FALSE, homdel = -2, hetdel = -0.2,
  dup = 0.2)
}
\arguments{
\item{data}{a data.frame representing a set of segments.}

\item{ind}{a character vector of length one. The sample (individual) for
which to extract the putative CNVs.}

\item{all_segments}{logical of length one (\code{TRUE} or \code{FALSE}).
Should all segments be extracted and output? Defaults to \code{FALSE},
meaning that only segments representing putative CNVs are output.}

\item{homdel}{a single numeric value. The threshold to be used for calling
homozygous deletions; all segments with a mean log2 ratio less than this
values is considered called as a homozygous deletion.}

\item{hetdel}{a single numeric value. The threshold to be used for calling
hemizygous deletions; all segments with a mean log2 ratio less than this
value and great than or equal to \code{homdel} is called as a hemizygous
deletion.}

\item{dup}{a single numeric value. The threshold to be used for calling
duplications; all segments with a mean log2 ratio larger than this value
is called as a duplication.}
}
\value{
a data.frame of the segments and associated metadata.
}
\description{
This function identifies a set of segments representing putative CNVs (
homozygous deletions, hemizygous deletions or duplications) from a set of
segments and associated mean log2 ratio values, and outputs them as a
\code{data.frame} along with some metadata describing them.
}
\examples{
NULL
}
