% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseg.R
\name{getseg}
\alias{getseg}
\title{Extract the segments from a given individual}
\usage{
getseg(data, ind)
}
\arguments{
\item{data}{a \code{data.frame} of mean log2 ratio values per bin, following
segmentation with function \code{\link{segment}}.}

\item{ind}{a character vector of length one. The individual for which the
data should be extracted (the name must correspond to one of the columns
of \code{data}).}
}
\value{
a \code{data.frame} with the segments of a given individual.
}
\description{
Extract the segments from a given individual
}
\examples{
NULL
}
