% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{gmax_binlist}
\alias{gmax_binlist}
\title{List of 1-kb bins for tallying reads in the soybean genome}
\format{A list of 20 numeric vectors. Elements of the list are named after
  the 20 chromosomes of the genome assembly, while the numeric vectors give
  the breakpoints of the 1-kb bins of each chromosome.}
\usage{
gmax_binlist
}
\description{
This list can be used as input to the function \code{\link{count_reads}} to
tally reads from a .bam file into 1-kb bins covering the whole assembly 2 of
the soybean (\emph{Glycine max}) genome.
}
\keyword{datasets}
