% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{gmax_chr_sizes}
\alias{gmax_chr_sizes}
\title{Length of the chromosomes of genome assembly 2 of soybean}
\format{A \code{data.frame} with 20 rows and 2 columns:
\describe{
  \item{chr}{character; the name of the chromosome in the assembly}
  \item{length}{integer; the number of base pairs of the chromosome}
}}
\source{
\url{http://plants.ensembl.org/Glycine_max/Info/Annotation/#assembly}
}
\usage{
gmax_chr_sizes
}
\description{
A dataset listing the lengths (in base pairs) of the 20 chromosomes of the
soybean (\emph{Glycine max}) genome assembly 2 (Glycine_max_v2.0)
}
\keyword{datasets}
