% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_counts.R
\name{merge_counts}
\alias{merge_counts}
\title{Merge two read count datasets}
\usage{
merge_counts(x, y)
}
\arguments{
\item{x, y}{the two read count \code{data.frame}'s to merge.}
}
\value{
a \code{data.frame} of read counts similar to those given as input,
  but that combines the data of the inputs.
}
\description{
This function takes two read count datasets as output by function
\code{link{count_reads}} and merges them, adding new rows and computing
the sum of read counts in shared bins as needed. The output is sorted
according to the genomic position of the bins, as the inputs were.
}
\examples{
NULL
}
