% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{var_diff}
\alias{var_diff}
\title{Difference based estimation of the variance}
\usage{
var_diff(x)
}
\arguments{
\item{x}{a numeric vector of log2 ratio values.}
}
\value{
a numeric vector of length one. The estimated variance of the log2
  ratios.
}
\description{
The value obtained from this function is used to set the parameter of the
segmentation (penalty and threshold).
}
\examples{
NULL
}
