% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{Segment the log2 ratios for a set of individuals}
\usage{
segment(data_matrix, info_indices, chrom_col, pos_col, threshold_param,
  lambda_param = 2, verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\description{
Segment the log2 ratios for a set of individuals
}
